refactor(#4196): split cqlite-cli/src/commands/salvage.rs into commands/salvage/{mod,discovery,report}.rs
salvage.rs crossed the ~800-line source threshold at round 14 (693 -> 833)
— since this PR created the file, the CQLITE_ALLOW_FILE_GROWTH=1 opt-out
(reserved for pre-existing files) does not apply, so splitting now rather
than carrying it under an opt-out (lead instruction).

Pure move, no behavior change: mod.rs keeps execute_salvage_command and
the module wiring; discovery.rs holds input-resolution/table-name
derivation (discover_salvage_inputs, table_name_from_input,
SkippedInput/SalvageDiscovery); report.rs holds the exit-code predicate,
table-dir-level finding attachment, and JSON/text manifest rendering
(report_is_imperfect, record_table_dir_level_findings,
exit_after_partial_failure, write_manifest_file, render_console) plus all
12 existing unit tests, unchanged. All three files individually well
under the ~800-line threshold (312 / 148 / 409 lines).

Verified --locked throughout (a read-only audit was reading the tree
concurrently): cargo check/clippy/test all clean, 12 unit tests + 9 CLI
integration tests pass unchanged, rustfmt --check clean on all 3 new
files with no sweep needed elsewhere.
